+++ src/components/policy/policiesTable.js
@@ -216,9 +216,6 @@ export default function PoliciesTable({ data, getData, access_modes, tenantName,
               indeterminate={numSelected > 0 && numSelected < rowCount}
               checked={rowCount > 0 && numSelected === rowCount}
               onChange={onSelectAllClick}
-              inputProps={{
-                'aria-label': 'select all desserts'
-              }}
             />
           </TableCell>
           {headCells.map((headCell) => (